Refactor: Move OAuth2Api struct & impls into oauth2.rs file.
